+++ CHANGES.md
@@ -1,3 +1,9 @@
+## Version 0.5.3
+
+- Functions to modify parameters of instanced network cells, conns and stims (issue #53)
+
+- Option to save figures with same name as data
+
 ## Version 0.5.2
 
 - Fixed bug that gave error since trying to import removed module default.py